Add methods to compute and get the degree of the polynomial
The method compute_degree checks if the ca_poly_t is proper. If that is the case, then degree + 1 is the length of the ca_poly_t. In this case, the method stores the degree in the degree property of the class. Otherwise, if the ca_poly_t isn't proper, the method doesn't compute the degree. The method get_degree then should only be called if compute_degree succeeded in computing the degree.